+++ js/jquery.gridy.js
@@ -540,16 +540,14 @@
 								number		= 0,
 								rangePage	= undefined,
 								start		= 1,
-								buttonMax	= self.opt.buttonMax,
+								buttonMax	= (self.opt.buttonPageNumber) ? self.opt.buttonMax : 0,
 								isEven		= (self.opt.buttonMax % 2 == 0);
 
-							if (self.opt.buttonMax > totalPage) {
+							if (buttonMax > totalPage) {
 								buttonMax = totalPage;
-							} else {
-								buttonMax = self.opt.buttonMax;
 							}
 
-							if (isEven) {
+							if (isEven && self.opt.buttonPageNumber) {
 								rangePage	= Math.ceil(buttonMax / 2);
 								start		= page - rangePage + 1;
 							} else {
@@ -582,12 +580,15 @@
 							}
 
 							var	hasExceeded			= totalPage > buttonMax,
-								hasBackNavigation	= hasExceeded && page > ((isEven) ? rangePage : rangePage + 1),
+								hasBackNavigation	= hasExceeded && page > ((isEven && self.opt.buttonPageNumber) ? rangePage : rangePage + 1),
 								hasNextNavigation	= hasExceeded && page < (totalPage - rangePage);
 
 							if (hasBackNavigation) {
 								buttons = '<input type="button" value="&lsaquo;" alt="' + self.opt.buttonBackTitle + '" title="' + self.opt.buttonBackTitle + '" class="gridy-back"/>&#160;';
-								buttons += buttonEmpty;
+								
+								if (self.opt.buttonPageNumber) {
+									buttons += buttonEmpty;
+								}
 							}
 
 							for (var i = start; i <= end; i++) {
@@ -596,13 +597,20 @@
 							}
 
 							if (hasNextNavigation) {
-								buttons += buttonEmpty;
-								buttons += '<input type="button" value="&rsaquo;" alt="' + self.opt.buttonNextTitle + '" title="' + self.opt.buttonNextTitle + '" class="gridy-button-next"/>&#160;';
+								if (self.opt.buttonPageNumber) {
+									buttons += buttonEmpty;
+								}
+
+								buttons += '<input type="button" value="&rsaquo;" alt="' + self.opt.buttonNextTitle + '" title="' + self.opt.buttonNextTitle + '" class="gridy-next"/>&#160;';
 							}
 
-							$buttons.html(buttons).children(':not(".gridy-button-back, .gridy-button-reticence, .gridy-button-next")').click(function() {
-								listGridy(parseInt(this.alt, 10), $currentSortName.val(), $currentSortOrder.val(), $this);
-							});
+							$buttons.html(buttons);
+
+							if (self.opt.buttonPageNumber) {
+								$buttons.children(':not(".gridy-back, .gridy-reticence, .gridy-next")').click(function() {
+									listGridy(parseInt(this.alt, 10), $currentSortName.val(), $currentSortOrder.val(), $this);
+								});
+							}
 
 							if (hasBackNavigation) {
 								$buttons.children('.gridy-back').click(function() {
@@ -611,7 +619,7 @@
 							}
 
 							if (hasNextNavigation) {
-								$buttons.children('.gridy-button-next').click(function() {
+								$buttons.children('.gridy-next').click(function() {
 									listGridy(page + 1, $currentSortName.val(), $currentSortOrder.val(), $this);
 								});
 							}
@@ -820,6 +828,7 @@
 		buttonMax			: 10,
 		buttonNextTitle		: 'Next &rsaquo;',
 		buttonOption		: true,
+		buttonPageNumber	: true,
 		buttonTitle			: 'page',
 		cache				: undefined,
 		clickFx				: false,